+++ quartic_bspline_extension/quartic_forward_cpu_kernel.cpp
@@ -2,57 +2,67 @@
 #include <torch/extension.h>
 #include <omp.h>
 
+#include "include/constants.h"
+#include "include/index_utils.h"
+#include "include/device_utils.h"
+
+/**
+ * @brief C++ kernel performing the forward step of the spline potential for
+ *      quartic (midpoint cardinal) b-spline.
+ * 
+ * @note
+ *  > Acts as fallback for quartic_bspline_forward_cuda_kernel() if no CUDA device
+ *      is detected.
+ *  > OpenMP-based implementation.
+ *  > Takes the same arguments as the corresponding CUDA kernel.
+ */
 template <typename T>
 void quartic_bspline_forward_cpu_kernel(
-    const torch::Tensor x,
+    const torch::PackedTensorAccessor32<T, 4> x,
     const torch::PackedTensorAccessor32<T, 2> weight_tensor,
     const torch::PackedTensorAccessor32<T, 1> centers,
     const T scale,
     const T scale_inv,
     const T delta_inv,
-    const size_t num_features,
-    const size_t width,
-    const size_t height,
-    torch::Tensor rho,
-    torch::Tensor rho_prime
+    torch::PackedTensorAccessor32<T, 4> rho,
+    torch::PackedTensorAccessor32<T, 4> rho_prime
 ){
-    T* x_ptr = x.data_ptr<T>();
-    T* rho_ptr = rho.data_ptr<T>();
-    T* rho_prime_ptr = rho_prime.data_ptr<T>();
+    #pragma omp parallel for collapse(4)
+    for (int64_t idx_bs = 0; idx_bs < x.size(0); idx_bs++) {
+        for (int64_t idx_f = 0; idx_f < x.size(1); idx_f++) {
+            for (int64_t idx_w = 0; idx_w < x.size(2); idx_w++) {
+                for (int64_t idx_h = 0; idx_h < x.size(3); idx_h++) {
+                    T rho_val = 0.0f;
+                    T rho_prime_val = 0.0f;
+                    
+                    const T x_ = x[idx_bs][idx_f][idx_w][idx_h];
+                    const std::pair<int, int> center_idx_bounds =
+                        compute_center_index_bounds(x_, centers[0], scale, delta_inv, centers.size(0));
 
-    #pragma omp parallel for
-    for (size_t i = 0; i < x.numel(); i ++){
-        T rho_val = 0.0f;
-        T rho_prime_val = 0.0f;
-        const T x_ = x_ptr[i];
+                    for (int j = center_idx_bounds.first; j <= center_idx_bounds.second; j++){
+                        const T x_scaled = (x_ - centers[j]) * scale_inv;
 
-        const std::pair<size_t, size_t> center_idx_bounds = ...;
+                        if (fabsf(x_scaled) < SUPP_RAD){
+                            int interval = (int)(x_scaled - SUPP_LOWER);
+                            interval = std::max(0, std::min(NUM_SUPP_INTERVALS - 1, interval));
 
-        for (size_t j = center_idx_bounds.first; j <= center_idx_bounds.second; j++){
-            const T x_scaled = (x_ - centers[j]) * scale_inv;
-
-            if (fabsf(x_scaled) < supp_rad){
-                int interval = (int)(x_scaled - supp_lower);
-                interval = std::max(0, std::min(num_supp_intervals - 1, interval));
-
-                // evaluate local spline and its derivative
-                T spline_val = quartic_bspline_coeffs[interval][4];
-                T spline_deriv = 0.0f;
-
-                #pragma unroll
-                for (auto i = k; k <= num_supp_intervals - 1; k++){
-                    spline_deriv = spline_deriv * x_scaled + spline_val;
-                    spline_val = spline_val * x_scaled 
-                               + quartic_bspline_coeffs[interval][num_supp_intervals - 1 - k];
+                            // evaluate local spline and its derivative
+                            T spline_val = QUARTIC_BSPLINE_COEFFS[interval][4];
+                            T spline_deriv = 0.0f;
+                            for (int k = 1; k <= NUM_SUPP_INTERVALS - 1; k++){
+                                spline_deriv = spline_deriv * x_scaled + spline_val;
+                                spline_val = spline_val * x_scaled 
+                                        + QUARTIC_BSPLINE_COEFFS[interval][NUM_SUPP_INTERVALS - 1 - k];
+                            }
+                            rho_val += weight_tensor[idx_f][j] * spline_val;
+                            rho_prime_val += weight_tensor[idx_f][j] * spline_deriv * scale_inv;                
+                        }
+                    }
+                    rho[idx_bs][idx_f][idx_w][idx_h] = rho_val;
+                    rho_prime[idx_bs][idx_f][idx_w][idx_h] = rho_prime_val;
                 }
-
-                size_t idx_f = (i / (width * height)) % num_features;
-                rho_val += weight_tensor[idx_f][j] * spline_val;
-                rho_prime_val += weight_tensor[idx_f][j] * spline_deriv * scale_inv;                
             }
         }
-        rho_ptr[i] = rho_val;
-        rho_prime_ptr[i] = rho_prime_val;
     }
 }
 
@@ -63,6 +73,8 @@ std::vector<torch::Tensor> quartic_bspline_forward_cpu_function(
     const torch::Tensor centers,
     const double scale
 ){
+    check_device_and_datatype({x, weight_tensor, centers});
+
     auto scalar_type = x.scalar_type();
 
     auto rho = torch::empty_like(x);
@@ -73,17 +85,14 @@ std::vector<torch::Tensor> quartic_bspline_forward_cpu_function(
 
     AT_DISPATCH_FLOATING_TYPES(scalar_type, "quartic_bspline_forward_cpu", [&] {
         quartic_bspline_forward_cpu_kernel<scalar_t>(
-            x,
+            x.packed_accessor32<scalar_t, 4>(),
             weight_tensor.packed_accessor32<scalar_t, 2>(), 
             centers.packed_accessor32<scalar_t, 1>(),
             static_cast<scalar_t>(scale),
             static_cast<scalar_t>(scale_inv),
             static_cast<scalar_t>(delta_inv),
-            x.size(1),
-            x.size(2),  // width
-            x.size(3),  // height
-            rho,
-            rho_prime
+            rho.packed_accessor32<scalar_t, 4>(),
+            rho_prime.packed_accessor32<scalar_t, 4>()
         );
     });