+++ tests/test_bspline_extension.py
@@ -1,29 +1,32 @@
 import torch
-from matplotlib import pyplot as plt
 import pytest
 
 from tests.QuarticBSplinePotential import QuarticBSplinePotential
-from bspline_cuda.functions import QuarticBSplineFunction
+from quartic_bspline_extension.functions import QuarticBSplineFunction
+
+dtypes = [torch.float64]
+devices = [torch.device('cpu')]
+if torch.cuda.is_available():
+    devices.append(torch.device('cuda:0'))
 
 @pytest.fixture(autouse=True)
 def seed_random_number_generators():
     seed_val = 123
     torch.manual_seed(seed_val)
 
-def test_forward():
-    device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-    dtype = torch.float64
-
+@pytest.mark.parametrize('device', devices)
+@pytest.mark.parametrize('dtype', dtypes)
+def test_forward(device: torch.device, dtype: torch.dtype) -> None:
     # NOTE
-    #   > For testing choose dtype = torch.float64
+    #   > For testing choose dtype = torch.float64; for other types torch.allclose()
+    #       with default tolerance may be too strict
     #   > For printing tensors with high precision use 
     #       torch.set_printoptions(precision=8) 
 
     bs = 10
-    f = 48
+    f = 4
     w = 64
     h = 128
-
     x = 5 * (2 * torch.rand(bs, f, w, h, device=device, dtype=dtype) - 1) 
     
     pot = QuarticBSplinePotential().to(device=device, dtype=dtype)
@@ -37,17 +40,17 @@ def test_forward():
 
     assert torch.allclose(y_true, y_test)
 
-def test_backward():
-    device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-    dtype = torch.float64
-
+@pytest.mark.parametrize('device', devices)
+@pytest.mark.parametrize('dtype', dtypes)
+def test_backward(device: torch.device, dtype: torch.dtype) -> None:
     bs = 10
-    f = 4
-    w = 128
-    h = 64
+    f = 2
+    w = 64
+    h = 128
 
     x = 5 * (2 * torch.rand(bs, f, w, h, device=device, dtype=dtype) - 1)
     x.requires_grad_(True)
+
     pot = QuarticBSplinePotential().to(device=device, dtype=dtype)
 
     weight_tensor = torch.cat([pot.weights.reshape(1, -1) for _ in range(0, f)], dim=0)
@@ -66,9 +69,8 @@ def test_backward():
 
     assert torch.allclose(dy_dx_true, dy_dx_test) and torch.allclose(dy_dw_true, dy_dw_test)
 
-def test_tensors_on_different_devices():
-    dtype = torch.float64
-
+@pytest.mark.parametrize('dtype', dtypes)
+def _test_tensors_on_different_devices(dtype: torch.dtype) -> None:
     device_1 = torch.device('cuda:0')
     device_2 = torch.device('cpu')
 
@@ -76,9 +78,9 @@ def test_tensors_on_different_devices():
     f = 4
     w = 64
     h = 32
-    num_centers = 33
-
     x = 5 * (2 * torch.rand(bs, f, w, h, device=device_1, dtype=dtype) - 1)
+
+    num_centers = 33
     weight_tensor = torch.rand(f, num_centers, device=device_1, dtype=dtype)
     box_lower = -3
     box_upper = 3
@@ -95,8 +97,8 @@ def test_tensors_on_different_devices():
     finally:
         assert expected_err_msg == err_msg
 
-def test_tensors_of_different_datatype():
-    device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
+@pytest.mark.parametrize('device', devices)
+def test_tensors_of_different_datatype(device: torch.device) -> None:
     dtype_1 = torch.float64
     dtype_2 = torch.float32
 
@@ -123,9 +125,9 @@ def test_tensors_of_different_datatype():
     finally:
         assert expected_err_msg == err_msg
 
-def test_vmap_forward():
-    device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-    dtype = torch.float64
+@pytest.mark.parametrize('device', devices)
+@pytest.mark.parametrize('dtype', dtypes)
+def test_vmap_forward(device: torch.device, dtype: torch.dtype):
 
     bs = 10
     f = 4
@@ -152,9 +154,9 @@ def test_vmap_forward():
 
     assert torch.allclose(func(*[x]), QuarticBSplineFunction.apply(x, weight_tensor, centers, scale)[0])
 
-def test_vmap_backward():
-    device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-    dtype = torch.float64
+@pytest.mark.parametrize('device', devices)
+@pytest.mark.parametrize('dtype', dtypes)
+def test_vmap_backward(device: torch.device, dtype: torch.dtype):
 
     bs = 10
     f = 4
@@ -187,5 +189,3 @@ def test_vmap_backward():
     dy_dx_test = torch.autograd.grad(inputs=x, outputs=y_test)[0]
 
     assert torch.allclose(dy_dx_true, dy_dx_test)
-
-